Fix get_bounds off-by-one(ish) error
In the existing code, we were always cropping at least one column,
because `lower=0` was impossible for `crop>0`: since the initial value
of `sum=0` will always be less than a non-zero threshold.

We want `lower` to be the first column where the total diff sum is
greater than the threshold, but because we incremented `lower` _after_
adding `diff_sums[lower]` to the total, we ended up with `lower`
representing the _second_ column after the total diff sum went over the
threshold.

In effect, this shifted all of our bounds one pixel away from the edge
of the image.