Keep grid width/height as floats
Since we're using this as a value to step by, the implicit truncation of
integer multiplication at that stage will result in the grid being
slightly less spaced out and favoring the left/top of the image.

Not a huge error, it's a maximum error of grid_size pixels, but that can
be significant for smaller images.

Since the bounds are inclusive, I also added 1 to the width, which is an
even smaller effect, but I think is more correct nonetheless.